Move header components to Layout folder and update imports accordingly
